remove CPR feature flag and replace old profile view
