+++ apps/betterangels/src/app/(private-screens)/_layout.tsx
@@ -162,26 +162,6 @@ export default function PrivateLayout() {
         }}
       />
       <Stack.Screen
-        name="add-client"
-        options={{
-          title: 'Add Client',
-          headerTitleAlign: 'center',
-          headerStyle: {
-            backgroundColor: Colors.BRAND_DARK_BLUE,
-          },
-          headerLeft: () => (
-            <TextButton
-              regular
-              color={Colors.WHITE}
-              fontSize="md"
-              accessibilityHint="goes to previous screen"
-              title="Back"
-              onPress={router.back}
-            />
-          ),
-        }}
-      />
-      <Stack.Screen
         name="edit-client/[id]"
         options={{
           title: 'Edit Client',
