--- apps/betterangels/src/app/(private-screens)/add-client.tsx
@@ -1,5 +1,0 @@
-import { AddEditClient } from '@monorepo/expo/betterangels';
-
-export default function AddClientScreen() {
-  return <AddEditClient />;
-}
