+++ libs/expo/betterangels/src/lib/providers/featureControls/constants.ts
@@ -1,5 +1,4 @@
 export const FeatureFlags = {
-  PROFILE_REDESIGN_FF: 'ffClientProfileRedesign',
   CLIENT_DEDUPE_FF: 'ffClientDedupe',
   APP_UPDATE_PROMPT_FF: 'ffAppUpdatePrompt',
   SHOW_DEBUG_INFO_FF: 'ffShowDebugInfo',
