--- libs/expo/betterangels/src/lib/screens/Client/Profile/types.ts
@@ -1,7 +1,0 @@
-import { ClientProfileQuery } from '../__generated__/Client.generated';
-
-export interface IProfileSectionProps {
-  expanded: undefined | string | null;
-  setExpanded: (expanded: undefined | string | null) => void;
-  client: ClientProfileQuery | undefined;
-}
