--- libs/expo/betterangels/src/lib/screens/Client/Schedule.tsx
@@ -1,10 +1,0 @@
-import { TextRegular } from '@monorepo/expo/shared/ui-components';
-import { ClientProfileQuery } from './__generated__/Client.generated';
-
-export default function Schedule({
-  client: _client,
-}: {
-  client: ClientProfileQuery | undefined;
-}) {
-  return <TextRegular>Schedule</TextRegular>;
-}
