--- libs/expo/betterangels/src/lib/screens/Client/Tasks.tsx
@@ -1,10 +1,0 @@
-import { TextRegular } from '@monorepo/expo/shared/ui-components';
-import { ClientProfileQuery } from './__generated__/Client.generated';
-
-export default function Tasks({
-  client: _client,
-}: {
-  client: ClientProfileQuery | undefined;
-}) {
-  return <TextRegular>Tasks</TextRegular>;
-}
