+++ libs/expo/betterangels/src/lib/screens/Client/index.tsx
@@ -9,40 +9,25 @@ import {
   ComponentType,
   ForwardRefExoticComponent,
   ReactElement,
-  RefObject,
   useEffect,
   useLayoutEffect,
-  useRef,
   useState,
 } from 'react';
 import { Pressable, View } from 'react-native';
-import { useFeatureFlagActive } from '../../hooks';
-import { FeatureFlags } from '../../providers';
 import { ClientProfileSectionEnum } from '../../screenRouting';
 import { MainContainer } from '../../ui-components';
 import ClientHeader from './ClientHeader';
-import ClientProfileView from './ClientProfile_V2';
+import ClientProfileView from './ClientProfile';
 import ClientTabs from './ClientTabs';
 import Docs from './Docs';
 import Interactions from './Interactions';
-import Locations from './Locations';
-import Profile from './Profile';
-import Schedule from './Schedule';
-import Services from './Services';
-import Tasks from './Tasks';
 import {
   ClientProfileQuery,
   useClientProfileQuery,
 } from './__generated__/Client.generated';
-interface ProfileRef {
-  scrollToRelevantContacts: () => void;
-}
-
 const getTabComponent = (
   key: string,
   client: ClientProfileQuery | undefined,
-  clientRedesignFeatureOn: boolean,
-  profileRef?: RefObject<ProfileRef>,
   openCard?: ClientProfileSectionEnum
 ): ReactElement | null => {
   const components: {
@@ -50,26 +35,19 @@ const getTabComponent = (
   } = {
     Docs,
     Interactions,
-    Locations,
-    Profile,
-    Schedule,
-    Services,
-    Tasks,
   };
 
-  const Component = components[key];
-
-  if (!Component) return null;
-
-  if (key !== 'Profile') {
-    return <Component client={client} />;
-  }
-
-  if (clientRedesignFeatureOn) {
+  if (key === 'Profile') {
     return <ClientProfileView client={client} openCard={openCard} />;
   }
 
-  return <Profile ref={profileRef} client={client} />;
+  const Component = components[key];
+
+  if (!Component) {
+    return null;
+  }
+
+  return <Component client={client} />;
 };
 
 export default function Client({
@@ -83,16 +61,6 @@ export default function Client({
 }) {
   const { data, loading, error } = useClientProfileQuery({ variables: { id } });
   const [tab, setTab] = useState('Profile');
-  const clientRedesignFeatureOn = useFeatureFlagActive(
-    FeatureFlags.PROFILE_REDESIGN_FF
-  );
-
-  const profileRef = useRef<ProfileRef | null>(null);
-
-  const handleScrollToRelevantContacts = async () => {
-    await setTab('Profile');
-    profileRef.current?.scrollToRelevantContacts();
-  };
 
   const navigation = useNavigation();
   const router = useRouter();
@@ -151,18 +119,9 @@ export default function Client({
 
   return (
     <MainContainer pt={0} pb={0} bg={Colors.NEUTRAL_EXTRA_LIGHT} px={0}>
-      <ClientHeader
-        onCaseManagerPress={handleScrollToRelevantContacts}
-        client={data?.clientProfile}
-      />
+      <ClientHeader client={data?.clientProfile} />
       <ClientTabs tab={tab} setTab={setTab} />
-      {getTabComponent(
-        tab,
-        data,
-        clientRedesignFeatureOn,
-        profileRef,
-        openCard
-      )}
+      {getTabComponent(tab, data, openCard)}
     </MainContainer>
   );
 }
